+++ docs/releases.md
@@ -5,6 +5,7 @@
 - [Github Sponsor page](https://github.com/sponsors/simulot)
 - [paypal donor page](https://www.paypal.com/donate/?hosted_button_id=VGU2SQE88T2T4)
 
+## Release 0.20.1
 
 ### changes
 - add git action to build and release